+++ docs/history.rst
@@ -4,11 +4,19 @@
 History
 =======
 
-Version 0.2.1 - Unreleased
+Version 0.2.2 - Unreleased
 --------------------------
 
-Fixed documentation build on RTFD. This won't be release as a new package version on
-PyPi since it is only related to RTFD.
+* Added ``smart_media.contrib.django_configuration.SmartMediaDefaultSettings`` class to
+  use with  `django-configuration <https://django-configurations.readthedocs.io/en/stable/>`_
+  to include default settings instead of ``from smart_media.settings import *``;
+
+
+
+Version 0.2.1 - 2023/01/11
+--------------------------
+
+Fixed documentation build on RTFD.
 
 
 Version 0.2.0 - 2023/01/11